+++ SE_Project/src/main/java/unisadventures/se_project/Model/basic_objects/DestructibleObject.java
@@ -10,7 +10,7 @@ public class DestructibleObject extends WorldObject implements DestructibleObjec
     private int _nextImageIndex;
     private static final int INTACT_OBJECT_IMAGE_FILE_NAME_INDEX = 0;
     
-    public DestructibleObject(int xPosition, int yPosition, int height, int width, ArrayList<String> imageFileNameList) {
+    public DestructibleObject(double xPosition, double yPosition, double height, double width, ArrayList<String> imageFileNameList) {
         super(xPosition, yPosition, height, width);
         _imageFileNameList = imageFileNameList;
         _listSizeMinusOne = _imageFileNameList.size() - 1;
